+++ src/main/java/com/example/campusfinder/board/request/entity/RequestBoard.java
@@ -1,6 +1,6 @@
 package com.example.campusfinder.board.request.entity;
 
-import com.example.campusfinder.board.student.entity.MeetingType;
+import com.example.campusfinder.board.coinstant.MeetingType;
 import com.example.campusfinder.core.base.BaseEntity;
 import com.example.campusfinder.home.entity.CategoryType;
 import jakarta.persistence.*;
